+++ app/src/main/java/com/luckyzyx/luckytool/hook/apps/launcher/FolderLayoutRowColume.kt
@@ -4,18 +4,23 @@ import com.highcapable.yukihookapi.hook.entity.YukiBaseHooker
 
 class FolderLayoutRowColume : YukiBaseHooker() {
     override fun onHook() {
-        findClass("com.android.launcher3.InvariantDeviceProfile").hook {
+        findClass("com.android.launcher3.folder.FolderGridOrganizer").hook {
             injectMember {
-                method {
-                    name = "initGrid"
+                constructor {
+                    paramCount = 1
                 }
                 afterHook {
-//                    field {
-//                        name = "numFolderRows"
-//                    }.get(instance).set(4)
+                    val x = field {
+                        name = "mMaxCountX"
+                    }.get(instance)
+                    val y = field {
+                        name = "mMaxCountY"
+                    }.get(instance)
+                    x.set(4)
+                    y.set(4)
                     field {
-                        name = "numFolderColumns"
-                    }.get(instance).set(4)
+                        name = "mMaxItemsPerPage"
+                    }.get(instance).set(x.cast<Int>()!! * y.cast<Int>()!!)
                 }
             }
         }