+++ app/src/main/java/com/luckyzyx/luckytool/hook/statusbar/StatusBarIcon.kt
@@ -37,7 +37,6 @@ class StatusBarIcon : YukiBaseHooker() {
                 loadHooker(RemoveMobileDataInout())
             }
 
-
         }
     }
 }