+++ app/src/main/java/com/luckyzyx/luckytool/utils/tools/DonateData.kt
@@ -11,6 +11,10 @@ object DonateData {
             add("加油加油 (10R)")
             add("天伞桜 (88.8R)")
             add("北风是不是冷 (15R)")
+            add("ssd.风格 (5R)")
+            add("挺好用的 (15R)")
+            add("松花蛋 (10R)")
+            add("佘樂 (20R)")
         }
     }
 }